Moving sharding methods into own file
