+++ src/jaxfun/galerkin/tensorproductspace.py
@@ -10,10 +10,17 @@ import jax
 import jax.numpy as jnp
 import sympy as sp
 from jax import Array, shard_map
-from jax.sharding import Mesh, NamedSharding, PartitionSpec as P
+from jax.sharding import NamedSharding, PartitionSpec as P
 
 from jaxfun.coordinates import CoordSys
-from jaxfun.typing import ArrayFun, MeshKind
+from jaxfun.sharding import (
+    _apply_separable_spmd_shard_map,
+    _build_local_apply_fn,
+    physical_sharding,
+    spectral_sharding,
+    spmd_mesh,
+)
+from jaxfun.typing import MeshKind
 from jaxfun.utils.common import jit_vmap, lambdify
 
 from .composite import BCGeneric, BoundaryConditions, Composite, DirectSum
@@ -68,18 +75,8 @@ class TensorProductSpace:
         self.name = name
         self.system: CoordSys = system
         self.tensorname = tensor_product_symbol.join([b.name for b in basespaces])
-        # Slab decomposition
-        self._spmd_mesh = Mesh(jax.devices(), ("k",))
-        # Sharding of arrays in spectral coefficient space.
-        self._spectral_sharding: NamedSharding | None = (
-            None if len(jax.devices()) == 1 else NamedSharding(self._spmd_mesh, P("k"))
-        )
-        # Sharding of arrays in physical space.
-        self._physical_sharding: NamedSharding | None = (
-            None
-            if len(jax.devices()) == 1
-            else NamedSharding(self._spmd_mesh, P(None, "k"))
-        )
+        self._spectral_sharding = spectral_sharding if len(jax.devices()) > 1 else None
+        self._physical_sharding = physical_sharding if len(jax.devices()) > 1 else None
         self._spmd_local_fn_cache: dict = {}
 
     def __len__(self) -> int:
@@ -237,15 +234,18 @@ class TensorProductSpace:
         cache_key = ("evaluate_mesh", kind, N)
         if cache_key not in self._spmd_local_fn_cache:
             self._spmd_local_fn_cache[cache_key] = tuple(
-                self._build_local_apply_fn(
+                _build_local_apply_fn(
+                    len(self),
                     ax,
                     partial(self.basespaces[ax].evaluate_mesh, kind=kind, N=N[ax]),
                 )
                 for ax in range(len(self))
             )
         fns = self._spmd_local_fn_cache[cache_key]
-        if self._spectral_sharding is not None and len(c.devices()) > 1:
-            return self._apply_separable_spmd_shard_map(c, fns, self._spectral_sharding)
+        if len(c.devices()) > 1:
+            return _apply_separable_spmd_shard_map(
+                c, fns, spectral_sharding, self._spmd_local_fn_cache
+            )
         for fn in fns:
             c = fn(c)
         return c
@@ -272,7 +272,7 @@ class TensorProductSpace:
         Returns:
             Scalar or (n_pts,) array of evaluated values.
         """
-        if self._spectral_sharding is not None and len(c.devices()) > 1:
+        if len(c.devices()) > 1:
             dim = len(self)
             T = self.basespaces
 
@@ -286,9 +286,8 @@ class TensorProductSpace:
                 dc = "abcdef"[:dim]
                 einsum_str = ",".join(f"j{ch}" for ch in dc) + f",{dc}->j"
 
-                c_spec = self._spectral_sharding.spec
-                mesh = self._spectral_sharding.mesh
-                physical_sharding = self._physical_sharding
+                c_spec = spectral_sharding.spec
+                mesh = spectral_sharding.mesh
 
                 def _local_eval(c_loc, C0_loc, *C_rest_loc):
                     return jax.lax.psum(
@@ -319,172 +318,6 @@ class TensorProductSpace:
             orthogonal_spaces, system=self.system, name=self.name + "o"
         )
 
-    def get_transposed_sharding(self, sharding: NamedSharding) -> NamedSharding:
-        """Return the sharding with unsharded and sharded axes transposed."""
-        if sharding is self._spectral_sharding:
-            return self._physical_sharding
-        elif sharding is self._physical_sharding:
-            return self._spectral_sharding
-        else:
-            raise ValueError("Provided sharding does not match spectral or physical.")
-
-    def _build_local_apply_fn(self, ax: int, fn: ArrayFun) -> ArrayFun:
-        """Return a ``jax.jit(jax.vmap(...))`` that applies *fn* along *ax*.
-
-        The resulting callable operates on a plain (non-sharded) local array,
-        so JAX compiles it once and reuses the compiled binary on every call.
-        """
-        dim = len(self)
-        if dim == 2:
-            axi = dim - 1 - ax
-            return jax.jit(jax.vmap(fn, in_axes=axi, out_axes=axi))
-        ax0, ax1 = sorted(set(range(dim)) - {ax})
-        return jax.jit(
-            jax.vmap(
-                jax.vmap(fn, in_axes=ax0, out_axes=ax0),
-                in_axes=ax1,
-                out_axes=ax1,
-            )
-        )
-
-    def _apply_separable_spmd(
-        self,
-        c: Array,
-        fns: tuple[ArrayFun, ...],
-        sharding: NamedSharding,
-    ) -> Array:
-        """Apply separable per-axis transforms on distributed (SPMD) arrays.
-
-        The transform is split into two fully-local phases separated by a
-        single all-to-all redistribution:
-
-        * **Phase 1 — unsharded axes**: each device holds the complete extent
-          along these axes, so no communication is needed.
-        * **All-to-all**: one ``jax.device_put`` transposes the sharding from
-          the originally-sharded axes to the formerly-unsharded axes.
-        * **Phase 2 — originally-sharded axes**: now fully local after the
-          transpose.
-
-        Note:
-        * The input sharding must be either spectral or physical, depending on
-          the transform being applied.
-        * The provided fns must be in the same order as basespaces and match the
-            sharding (e.g. spectral fns applied with spectral sharding).
-        * When input sharding is spectral, the output is physical and vice versa.
-
-        """
-        dim = len(self)
-        spec = sharding.spec
-        sharded = [ax for ax in range(dim) if ax < len(spec) and spec[ax] is not None]
-        unsharded = [ax for ax in range(dim) if ax not in sharded]
-        n_local = jax.local_device_count()
-
-        # Phase 1 — unsharded axes: operate on each local addressable shard.
-        # fns[ax] is a pre-jitted vmap; XLA cache is hit on every call.
-        local_shards = [c.addressable_data(d) for d in range(n_local)]
-        for ax in unsharded:
-            local_shards = [fns[ax](shard) for shard in local_shards]
-
-        # Reconstruct the global array from the updated local shards.
-        # Unsharded axes may have changed size (e.g. Chebyshev with BCs);
-        # sharded axes retain their original global size.
-        global_shape_p1 = tuple(
-            local_shards[0].shape[ax] if ax in unsharded else c.shape[ax]
-            for ax in range(dim)
-        )
-        c = jax.make_array_from_single_device_arrays(
-            global_shape_p1, sharding, local_shards
-        )
-
-        # All-to-all: transpose the sharding (one collective, O(N^d/P) per device).
-        # The two pre-built shardings are each other's transpose by construction.
-        transposed = self.get_transposed_sharding(sharding)
-
-        c = jax.device_put(c, transposed)
-
-        # Phase 2 — originally-sharded axes: now fully local after the transpose.
-        local_shards = [c.addressable_data(d) for d in range(n_local)]
-        for ax in sharded:
-            local_shards = [fns[ax](shard) for shard in local_shards]
-
-        # Reconstruct the final global array; sharded-axis sizes may have changed.
-        global_shape_p2 = list(global_shape_p1)
-        for ax in sharded:
-            global_shape_p2[ax] = local_shards[0].shape[ax]
-        return jax.make_array_from_single_device_arrays(
-            tuple(global_shape_p2), transposed, local_shards
-        )
-
-    def _apply_separable_spmd_shard_map(
-        self,
-        c: Array,
-        fns: tuple[ArrayFun, ...],
-        sharding: NamedSharding,
-    ) -> Array:
-        """Apply separable per-axis transforms using ``shard_map`` + ``lax.all_to_all``.
-
-        JAX-native alternative to :meth:`_apply_separable_spmd`.  The entire
-        transform — including the inter-device redistribution — is a single
-        compiled XLA computation, allowing XLA to fuse across phase boundaries.
-
-        The algorithm mirrors the three-phase structure of the addressable-data
-        approach:
-
-        * **Phase 1**: unsharded-axis transforms applied locally inside the kernel.
-        * **All-to-all**: ``lax.all_to_all(tiled=True)`` transposes the sharding.
-        * **Phase 2**: originally-sharded-axis transforms applied locally.
-
-        .. note::
-            ``lax.all_to_all(tiled=True)`` requires the ``split_axis`` dimension
-            (the first unsharded axis, after Phase 1) to be divisible by the
-            total number of devices.  This holds for typical spectral sizes
-            (powers of two for Fourier, even quadrature counts for Chebyshev).
-
-        """
-        # Cache the compiled shard_map function keyed on the (fns, sharding spec)
-        # combination.  _kernel is defined inside the method, so each call would
-        # produce a new function object and force recompilation.  Storing the
-        # shard_map-wrapped callable ensures it is compiled exactly once.
-        cache_key = ("shard_map_kernel", id(fns), sharding.spec)
-        if cache_key not in self._spmd_local_fn_cache:
-            dim = len(self)
-            spec = sharding.spec
-            sharded = [
-                ax for ax in range(dim) if ax < len(spec) and spec[ax] is not None
-            ]
-            unsharded = [ax for ax in range(dim) if ax not in sharded]
-
-            transposed = self.get_transposed_sharding(sharding)
-
-            def _kernel(c_loc: Array) -> Array:
-                # Phase 1 — unsharded axes: fully local, no communication.
-                for ax in unsharded:
-                    c_loc = fns[ax](c_loc)
-                # All-to-all: redistribute sharding from sharded → unsharded axes.
-                c_loc = jax.lax.all_to_all(
-                    c_loc,
-                    axis_name="k",
-                    split_axis=unsharded[0],
-                    concat_axis=sharded[0],
-                    tiled=True,
-                )
-                # Phase 2 — originally-sharded axes: fully local after the transpose.
-                for ax in sharded:
-                    c_loc = fns[ax](c_loc)
-                return c_loc
-
-            self._spmd_local_fn_cache[cache_key] = jax.jit(
-                shard_map(
-                    _kernel,
-                    mesh=sharding.mesh,
-                    in_specs=(sharding.spec,),
-                    out_specs=transposed.spec,
-                    check_vma=False,
-                )
-            )
-
-        return self._spmd_local_fn_cache[cache_key](c)
-
     def backward(
         self,
         c: Array,
@@ -502,15 +335,18 @@ class TensorProductSpace:
         cache_key = ("backward", N)
         if cache_key not in self._spmd_local_fn_cache:
             self._spmd_local_fn_cache[cache_key] = tuple(
-                self._build_local_apply_fn(
+                _build_local_apply_fn(
+                    len(self),
                     ax,
                     partial(self.basespaces[ax].backward, N=N[ax]),
                 )
                 for ax in range(len(self))
             )
         fns = self._spmd_local_fn_cache[cache_key]
-        if self._spectral_sharding is not None and len(c.devices()) > 1:
-            return self._apply_separable_spmd_shard_map(c, fns, self._spectral_sharding)
+        if self._spectral_sharding and len(c.devices()) > 1:
+            return _apply_separable_spmd_shard_map(
+                c, fns, spectral_sharding, self._spmd_local_fn_cache
+            )
         for fn in fns:
             c = fn(c)
         return c
@@ -524,12 +360,14 @@ class TensorProductSpace:
         cache_key = ("scalar_product",)
         if cache_key not in self._spmd_local_fn_cache:
             self._spmd_local_fn_cache[cache_key] = tuple(
-                self._build_local_apply_fn(ax, self.basespaces[ax].scalar_product)
+                _build_local_apply_fn(len(self), ax, self.basespaces[ax].scalar_product)
                 for ax in range(len(self))
             )
         fns = self._spmd_local_fn_cache[cache_key]
-        if self._physical_sharding is not None and len(u.devices()) > 1:
-            return self._apply_separable_spmd_shard_map(u, fns, self._physical_sharding)
+        if self._physical_sharding and len(u.devices()) > 1:
+            return _apply_separable_spmd_shard_map(
+                u, fns, physical_sharding, self._spmd_local_fn_cache
+            )
         for fn in fns:
             u = fn(u)
         return u
@@ -539,12 +377,14 @@ class TensorProductSpace:
         cache_key = ("forward",)
         if cache_key not in self._spmd_local_fn_cache:
             self._spmd_local_fn_cache[cache_key] = tuple(
-                self._build_local_apply_fn(ax, self.basespaces[ax].forward)
+                _build_local_apply_fn(len(self), ax, self.basespaces[ax].forward)
                 for ax in range(len(self))
             )
         fns = self._spmd_local_fn_cache[cache_key]
-        if self._physical_sharding is not None and len(u.devices()) > 1:
-            return self._apply_separable_spmd_shard_map(u, fns, self._physical_sharding)
+        if self._physical_sharding and len(u.devices()) > 1:
+            return _apply_separable_spmd_shard_map(
+                u, fns, physical_sharding, self._spmd_local_fn_cache
+            )
         for fn in fns:
             u = fn(u)
         return u
@@ -563,7 +403,8 @@ class TensorProductSpace:
         cache_key = ("backward_primitive", k, N)
         if cache_key not in self._spmd_local_fn_cache:
             self._spmd_local_fn_cache[cache_key] = tuple(
-                self._build_local_apply_fn(
+                _build_local_apply_fn(
+                    len(self),
                     ax,
                     partial(
                         self.basespaces[ax].backward_primitive,
@@ -574,8 +415,10 @@ class TensorProductSpace:
                 for ax in range(len(self))
             )
         fns = self._spmd_local_fn_cache[cache_key]
-        if self._spectral_sharding is not None and len(c.devices()) > 1:
-            return self._apply_separable_spmd_shard_map(c, fns, self._spectral_sharding)
+        if self._spectral_sharding and len(c.devices()) > 1:
+            return _apply_separable_spmd_shard_map(
+                c, fns, spectral_sharding, self._spmd_local_fn_cache
+            )
         for fn in fns:
             c = fn(c)
         return c
@@ -583,18 +426,18 @@ class TensorProductSpace:
     def to_orthogonal(self, c: Array) -> Array:
         """Return coefficients c mapped to underlying orthogonal basis."""
         dim = len(self)
-        if self._spectral_sharding is not None and len(c.devices()) > 1:
+        if self._spectral_sharding and len(c.devices()) > 1:
             S = [s.S for s in self.basespaces]
             cache_key = "to_orthogonal"
             if cache_key not in self._spmd_local_fn_cache:
                 self._spmd_local_fn_cache[cache_key] = tuple(
-                    self._build_local_apply_fn(ax, S[ax].rmatvec) for ax in range(dim)
+                    _build_local_apply_fn(dim, ax, S[ax].rmatvec) for ax in range(dim)
                 )
             fns = self._spmd_local_fn_cache[cache_key]
-            result = self._apply_separable_spmd_shard_map(
-                c, fns, self._spectral_sharding
+            result = _apply_separable_spmd_shard_map(
+                c, fns, spectral_sharding, self._spmd_local_fn_cache
             )
-            return jax.device_put(result, self._spectral_sharding)
+            return jax.device_put(result, spectral_sharding)
         if dim == 2:
             S = [s.S for s in self.basespaces]
             return S[0].T @ c @ S[1]
@@ -605,16 +448,18 @@ class TensorProductSpace:
         """Return coefficients c mapped from underlying orthogonal basis."""
         S = [s.get_inverse_stencil() for s in self.basespaces]
         dim = len(self)
-        if self._physical_sharding is not None and len(c.devices()) > 1:
+        if self._physical_sharding and len(c.devices()) > 1:
             cache_key = "from_orthogonal"
             if cache_key not in self._spmd_local_fn_cache:
                 self._spmd_local_fn_cache[cache_key] = tuple(
-                    self._build_local_apply_fn(ax, lambda x, _S=S[ax]: jnp.dot(x, _S))
+                    _build_local_apply_fn(dim, ax, lambda x, _S=S[ax]: jnp.dot(x, _S))
                     for ax in range(dim)
                 )
             fns = self._spmd_local_fn_cache[cache_key]
             c = jax.device_put(c, self._physical_sharding)
-            return self._apply_separable_spmd_shard_map(c, fns, self._physical_sharding)
+            return _apply_separable_spmd_shard_map(
+                c, fns, self._physical_sharding, self._spmd_local_fn_cache
+            )
         if dim == 2:
             return S[0].T @ c @ S[1]
         return jnp.einsum("is,jp,kl,ijk->spl", *S, c)
@@ -653,17 +498,14 @@ class VectorTensorProductSpace:
         self.num_quad_points = self.tensorspaces[0].num_quad_points
         # Slab decomposition for vector spaces
         # First index is vector component, which is not sharded.
-        self._spmd_mesh = Mesh(jax.devices(), ("k",))
         self._spectral_sharding: NamedSharding | None = (
-            None
-            if len(jax.devices()) == 1
-            else NamedSharding(self._spmd_mesh, P(None, "k"))
+            None if len(jax.devices()) == 1 else NamedSharding(spmd_mesh, P(None, "k"))
         )
         # Sharding of arrays in physical space.
         self._physical_sharding: NamedSharding | None = (
             None
             if len(jax.devices()) == 1
-            else NamedSharding(self._spmd_mesh, P(None, None, "k"))
+            else NamedSharding(spmd_mesh, P(None, None, "k"))
         )
 
     def __len__(self) -> int:
@@ -890,6 +732,8 @@ class DirectSumTPS(TensorProductSpace):
         self.name = name
         self.bndvals: dict[tuple[OrthogonalSpace, ...], Array] = {}
         self.tensorname = tensor_product_symbol.join([b.name for b in basespaces])
+        self._spectral_sharding = spectral_sharding if len(jax.devices()) > 1 else None
+        self._physical_sharding = physical_sharding if len(jax.devices()) > 1 else None
 
         # Normalize symbolic BC expressions to base scalar form
         for space in basespaces:
@@ -986,14 +830,6 @@ class DirectSumTPS(TensorProductSpace):
                 else:
                     self.bndvals[tensorspace] = jnp.array(uh).T
         self.orthogonal = self.get_orthogonal()
-
-    @property
-    def _physical_sharding(self) -> NamedSharding | None:
-        return self.tpspaces[next(iter(self.tpspaces))]._physical_sharding
-
-    @property
-    def _spectral_sharding(self) -> NamedSharding | None:
-        return self.tpspaces[next(iter(self.tpspaces))]._spectral_sharding
 
     def split(
         self, spaces: list[OrthogonalSpace | DirectSum]
