+++ tests/galerkin/test_forward_backward_spmd.py
@@ -24,6 +24,7 @@ from jaxfun.galerkin import (
     TensorProduct,
 )
 from jaxfun.galerkin.inner import project
+from jaxfun.sharding import physical_sharding, spectral_sharding
 from jaxfun.utils.common import ulp
 
 pytestmark = pytest.mark.spmd
@@ -67,9 +68,9 @@ def test_forward_backward_2d_spmd(space0, space1) -> None:
 )
 def test_scalar_product_2d_spmd(space0, space1) -> None:
     T = TensorProduct(space0(8), space1(8))
-    u = jax.device_put(jnp.ones(T.shape()), T._physical_sharding)
+    u = jax.device_put(jnp.ones(T.shape()), physical_sharding)
     uh = T.scalar_product(u)
-    assert uh.sharding == T._spectral_sharding
+    assert uh.sharding == spectral_sharding
 
 
 # ---------------------------------------------------------------------------
@@ -106,9 +107,9 @@ def test_forward_backward_3d_spmd(space0, space1, space2) -> None:
 )
 def test_scalar_product_3d_spmd(space0, space1, space2) -> None:
     T = TensorProduct(space0(8), space1(8), space2(8))
-    u = jax.device_put(jnp.ones(T.shape()), T._physical_sharding)
+    u = jax.device_put(jnp.ones(T.shape()), physical_sharding)
     uh = T.scalar_product(u)
-    assert uh.sharding == T._spectral_sharding
+    assert uh.sharding == spectral_sharding
 
 
 @pytest.mark.parametrize("domain", [(-1, 1), (0, 2), (-2, 2)])
